+++ src/main/java/org/sonar/plugins/delphi/pmd/rules/LowerCaseReservedWordsRule.java
@@ -26,12 +26,6 @@ public class LowerCaseReservedWordsRule extends AbstractDelphiRule {
     if (!DelphiKeywords.KEYWORDS.contains(node.jjtGetId())) {
       return false;
     }
-
-    if (DelphiKeywords.SPECIAL_KEYWORDS.contains(node.jjtGetId())) {
-      // Let SpecialKeywordCapitalizationRule handle these
-      return false;
-    }
-
     return !StringUtils.isAllLowerCase(node.getToken().getImage());
   }
 }